diplay distinct w densities as distinct graphs, allows large minimizers

+++ src/lib.rs
@@ -5,7 +5,7 @@ pub mod superkmer;
 pub mod superkmers_computation;
 mod two_bits;
 
-type Minimizer = u64;
+type Minimizer = u128;
 
 pub use crate::superkmers_computation::compute_all_superkmers_linear_streaming;
 pub use crate::superkmers_computation::compute_superkmers_linear_streaming;
